pyproject.toml
[project]
name = "bsf-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lancedb>=0.20.0",
    "nicegui>=2.11.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "sentence-transformers>=3.4.1",
    "tantivy>=0.22.0",
]


